fix(ui): avoid a narrowing-conversion warning in pairPortsByOrder
QList::size() returns qsizetype (64-bit); std::min() of two of those
implicitly narrowed into an int, which GCC's -Werror=conversion (enabled
in CI's ubsan/asan builds) correctly flags -- MSVC's warning set didn't
catch it locally. Make the truncation explicit with static_cast<int>.

+++ App/UI/ElementContextMenu.cpp
@@ -58,7 +58,7 @@ QList<OutputPort *> allOutputPorts(GraphicElement *elm)
 QList<PortPair> pairPortsByOrder(const QList<OutputPort *> &outputs, const QList<InputPort *> &inputs)
 {
     QList<PortPair> result;
-    const int count = std::min(outputs.size(), inputs.size());
+    const int count = static_cast<int>(std::min(outputs.size(), inputs.size()));
     result.reserve(count);
 
     for (int i = 0; i < count; ++i) {
